refactor: move GamePlay and lower components
게임플레이를 컴포넌트 폴더로 이동하고 그 하위 폴더도 게임플레이 안으로 이동
