+++ front/src/components/GamePlay/CanvasSection/CanvasSection.style.js
