+++ front/src/components/GamePlay/CanvasSection/DrawingPlayGround/DrawingPlayGround.js
@@ -1,8 +1,8 @@
 import React, { useReducer, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import Tools from 'components/CanvasSection/DrawingPlayGround/Tools/Tools';
-import PainterBoard from 'components/CanvasSection/DrawingPlayGround/PainterBoard/PainterBoard';
-import NonPainterBoard from 'components/CanvasSection/DrawingPlayGround/NonPainterBoard/NonPainterBoard';
+import Tools from 'components/GamePlay/CanvasSection/DrawingPlayGround/Tools/Tools';
+import PainterBoard from 'components/GamePlay/CanvasSection/DrawingPlayGround/PainterBoard/PainterBoard';
+import NonPainterBoard from 'components/GamePlay/CanvasSection/DrawingPlayGround/NonPainterBoard/NonPainterBoard';
 import DrawingPlayGroundStyle from './DrawingPlayGround.style';
 
 DrawingPlayGround.propTypes = {
