+++ front/src/components/GamePlay/CanvasSection/DrawingPlayGround/NonPainterBoard/NonPainterBoard.style.js
