+++ front/src/components/GamePlay/CanvasSection/DrawingPlayGround/Tools/ColorPicker/ColorBox.style.js
