+++ front/src/components/GamePlay/CanvasSection/DrawingPlayGround/Tools/ColorPicker/ColorPicker.style.js
