+++ front/src/components/GamePlay/CanvasSection/Section.style.js
