+++ front/src/components/GamePlay/CanvasSection/WordChoice/Background.style.js
