+++ front/src/components/GamePlay/CanvasSection/WordChoice/WordCard.style.js
