+++ front/src/components/GamePlay/CanvasSection/WordChoice/WordChoice.js
