+++ front/src/components/GamePlay/GameLoading/GameLoading.js
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
+import GamePlayContext from 'components/GamePlay/GamePlay.context';
 import FullScreen from './GameLoading.style';
-import Loading from '../globalComponents/Loading/Loading';
-import GamePlayContext from 'pages/GamePlay/GamePlay.context';
+import Loading from '../../globalComponents/Loading/Loading';
 
 export default function GameLoading() {
   const { userList } = useContext(GamePlayContext);
