+++ front/src/components/GamePlay/GameLoading/GameLoading.style.js
