+++ front/src/pages/SecretGame/SecretGame.js
@@ -1,8 +1,8 @@
 import React, { useEffect, useContext, useState } from 'react';
 import GlobalContext from 'global.context';
 import SecretGameContext from 'pages/SecretGame/SecretGame.context';
-import GamePlay from 'pages/GamePlay/GamePlay';
-import GamePlayContext from 'pages/GamePlay/GamePlay.context';
+import GamePlay from 'components/GamePlay/GamePlay';
+import GamePlayContext from 'components/GamePlay/GamePlay.context';
 
 import GamePreparation from 'pages/SecretGame/GamePreparation';
 import Room from 'logics/room/index';
